+++ Changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.8.1] - 11-12-2020
+
+Requires Elpi 1.12 and Coq 8.13.
+### HOAS
+- Illformed terms like `global (const X)` (which have no
+  representation in Coq) are now reported with a proper error message.
+  Whe passed to `coq.term->string`, instead of a fatal error, we pick for
+  the illformed sub term the `unknown_gref` special constant.
+
 ## [1.8.0] - 29-11-2020
 
 Requires Elpi 1.12 and Coq 8.12.